Restructure config source fields for better readability
