+++ Model/Config/Source/CaptchaType.php
@@ -5,13 +5,13 @@ namespace Smaily\SmailyForMagento\Model\Config\Source;
 class CaptchaType implements \Magento\Framework\Option\ArrayInterface
 {
     /**
-     * Get Option values for CAPTCHA type field.
+     * Get options for CAPTCHA type field.
      *
      * @return array
      */
     public function toOptionArray()
     {
-        $list = [
+        return [
             [
                 'value' => 'magento_captcha',
                 'label' => 'Magento Built-In CAPTCHA',
@@ -21,7 +21,5 @@ class CaptchaType implements \Magento\Framework\Option\ArrayInterface
                 'label' => 'Google reCAPTCHA',
             ],
         ];
-
-        return $list;
     }
 }
